+++ finance/urls.py
@@ -7,5 +7,6 @@ urlpatterns = [
     path("estimate/<e_id>", views.viewEstimate.as_view(), name="estimate"),
     path("invoice/<e_id>", views.viewInvoice.as_view(), name="invoice"),
     path("timesheet/<pp_id>/<emp_id>", views.viewTimesheet.as_view(), name="timesheet"),
+    path("rollover/<pp_id>/<emp_id>", views.RollOverAllShifts.as_view(), name="rollover"),
     path("summary/<pp_id>/", views.viewSummary.as_view(), name="summary"),
 ]